Updating DoneButton and moving the component to buttons folder
